#503: Fix backend price precision issue

+++ Api/Xipona.Api.Contracts/Items/Commands/UpdateItemPrice/UpdateItemPriceContract.cs
@@ -4,7 +4,7 @@ namespace ProjectHermes.Xipona.Api.Contracts.Items.Commands.UpdateItemPrice
 {
     public class UpdateItemPriceContract
     {
-        public UpdateItemPriceContract(Guid? itemTypeId, Guid storeId, float price)
+        public UpdateItemPriceContract(Guid? itemTypeId, Guid storeId, decimal price)
         {
             ItemTypeId = itemTypeId;
             StoreId = storeId;
@@ -13,6 +13,6 @@ namespace ProjectHermes.Xipona.Api.Contracts.Items.Commands.UpdateItemPrice
 
         public Guid? ItemTypeId { get; set; }
         public Guid StoreId { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
