+++ Frontend/Xipona.Frontend.Redux/ShoppingList/Actions/PriceUpdater/PriceOnPriceUpdaterChangedAction.cs
@@ -1,2 +1,2 @@
 ﻿namespace ProjectHermes.Xipona.Frontend.Redux.ShoppingList.Actions.PriceUpdater;
-public record PriceOnPriceUpdaterChangedAction(float Price);
+public record PriceOnPriceUpdaterChangedAction(decimal Price);
